Update build and modify permission check function
