+++ client/src/js/references/components/Detail/Detail.js
@@ -5,9 +5,9 @@ import { Switch, Route, Redirect } from "react-router-dom";
 import { push } from "react-router-redux";
 import { LinkContainer } from "react-router-bootstrap";
 import { Badge, Nav, NavItem } from "react-bootstrap";
-import { filter, find } from "lodash-es";
 import { getReference } from "../../actions";
 import { LoadingPlaceholder, Icon, ViewHeader, Flex, FlexItem } from "../../../base";
+import { checkUserRefPermission } from "../../../utils";
 
 import EditReference from "./Edit";
 import ReferenceManage from "./Manage";
@@ -17,34 +17,6 @@ import ReferenceOTUs from "../../../otus/components/List";
 import ReferenceIndexList from "../../../indexes/components/List";
 import SourceTypes from "../../../administration/components/General/SourceTypes";
 import InternalControl from "../../../administration/components/General/InternalControl";
-
-const checkUserModifyRefPermission = (props) => {
-    const { isAdmin, userId, userGroups } = props;
-    const refUsers = props.detail.users;
-    const refGroups = props.detail.groups;
-
-    if (isAdmin) {
-        return true;
-    }
-
-    const userExists = find(refUsers, ["id", userId]);
-
-    const groupsExist = filter(refGroups, refGroup => {
-        const result = filter(userGroups, group => group === refGroup.id);
-        return result;
-    });
-
-    if (userExists && userExists.modify) {
-        return true;
-    }
-
-    if (groupsExist.length) {
-        const hasBuildPermission = filter(groupsExist, group => group.modify);
-        return !!hasBuildPermission.length;
-    }
-
-    return false;
-};
 
 const ReferenceSettings = ({ isRemote }) => (
     <div className="settings-container">
@@ -68,8 +40,7 @@ class ReferenceDetail extends React.Component {
         }
 
         const { name, id, remotes_from, created_at, user } = this.props.detail;
-
-        const hasModify = checkUserModifyRefPermission(this.props);
+        const hasModify = checkUserRefPermission(this.props, "modify");
 
         let headerIcon;
